fix: organize code into a yaml processor
